requirements.txt
nltk==3.7
pandas==1.4.3
praw==7.6.0
python-decouple==3.6
tweepy==4.10.0
